Grab both header and body
The entries are split into two parts, a header with the time and
information about when things happened and then a body which has all the
details. Both are handy, but splitting based on the header separator is
tricky, so the split splits the header and body into two different
pices, and then they need to be recombined later.