Fix regexes for parsing type and address
Address was looking for a space to end, but addresses have spaces in
them! Oops. They also don't always end with a comma. Sometimes there is
a GRID entry in there, but sometimes there is not. We found the most
reliable way is to get everything before Clearance Code, because that
will reliably be there.