Correct imporperly parsed log entries
Split out slopd pulling and parsing

I want to be able to reparse the slopd logs that are already in
the database to correct old entries that had bad parsing rules.
Split up parsing and pulling will allow me to use the raw field
in the DB to reparse and then update the improperly parsed logs.