Fixup the Dockerfile and dependency management of python
The dependency management was using requirements, which isn't as nice as
Pipfile, so remove it.

Docker was busting the cache of installing system packages through apt,
so it was doing this step every single time which was troublesome for
speed. Placing the dependency management files after system packages
lets the system packages always be cached, and then then adding in all
code only after dependencies are installed allows the dependency layer
to stay cached as well ubnless they are actually updated.