Add a few changes to the kubernetes conf
Added a few dev resources for lazy creation of a dev env. The better way
to do this would be to template some things out so there isn't so much
duplication, but that will have to wait for a bit later.

Also configured the Digital Ocean Load Balancers a bit more so they now
will actually do HTTPS correctly.