Fix kafka service type and improve logging and build script
The kafka service had a load balancer, which was not helpful for
accessing kafka within the cluster. This is important that we don't
expose this externally, so I destroyed all of the old Kafka and
zookeeper bits just to make sure nobody put anything on them and then
remade them with the correct ClusterIP default so it isn't exposed.

Added some more logging around insertion into the kafka broker, and
adding IDs to log messages for observability.

Added more to the deploy script to run an instance of the cronjob to
test it out.