Clean up kubernetes resources and fix service selector bug
The selector bug was caused by giving every resource the same app name.
The selector for the service was using the app name, but apparently
common labels will override selector as well as the metadata labels for
the service as well. So the service was grabbing every single resource
it could since everything was named crime-map, and the selector was
changed from mongodb to crime-map since that was also apparently a label
to be updated. This was not how I expected kustomize to behave, but for
some reason that's what happened.

There were also a bunch of dev specific resources that needed to be
deleted or moved in as resources to the kustomize overlays for the
specific environments that aren't dev or integration.