Ensure that saving a new alert doesn't create new ones
When saving a newly created alert clientside inside the UI, it would
continuously create new alerts since the client wasn't maintaining the
ObjectId that was created by mongodb. Now this will pass the ObjectId
back to the client and the client will update the alert accordingly so
that only one is created.